addendum - addressing latest comments
